Move clap args stuff into it's own file: args.rs.
